zbyszek work - cleanup
