--- src/DomainModel/Repository/EarthResearchStationRepository.php
@@ -1,8 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace App\DomainModel\Repository;
-
-interface EarthResearchStationRepository
-{
-}
