--- src/DomainModel/Repository/ProductRepository.php
@@ -1,8 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace App\DomainModel\Repository;
-
-interface ProductRepository
-{
-}
