+++ src/Entity/Event.php
@@ -3,11 +3,10 @@ declare(strict_types=1);
 
 namespace App\Entity;
 
-use App\Repository\EventRepository;
 use Doctrine\ORM\Mapping as ORM;
 
 /**
- * @ORM\Entity(repositoryClass=EventRepository::class)
+ * @ORM\Entity()
  */
 class Event
 {
