+++ src/Expeditions/Application/Handler/FindExpeditionByIdQueryHandler.php
@@ -3,9 +3,9 @@ declare(strict_types=1);
 
 namespace App\Expeditions\Application\Handler;
 
-use App\DomainModel\Repository\ExpeditionRepository;
 use App\Entity\Expedition;
 use App\Expeditions\Application\Query\FindExpeditionByIdQuery;
+use App\Expeditions\Domain\ExpeditionRepository;
 use Symfony\Component\Messenger\Handler\MessageHandlerInterface;
 
 class FindExpeditionByIdQueryHandler implements MessageHandlerInterface
