+++ src/Expeditions/Domain/UserRepository.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace App\DomainModel\Repository;
+namespace App\Expeditions\Domain;
 
 use App\Entity\User;
