+++ src/Expeditions/Framework/Controller.php
@@ -1,8 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Presentation;
+namespace App\Expeditions\Framework;
 
+use App\Shared\Application\Command;
 use App\Shared\Application\Query;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Messenger\Envelope;
@@ -10,18 +11,30 @@ use Symfony\Component\Messenger\MessageBusInterface;
 use Symfony\Component\Messenger\Stamp\DelayStamp;
 use Symfony\Component\Messenger\Stamp\HandledStamp;
 
-abstract class QueryController extends AbstractController
+abstract class Controller extends AbstractController
 {
-    private MessageBusInterface $queryBus;
+    private MessageBusInterface $bus;
 
-    public function __construct(MessageBusInterface $queryBus)
+    public function __construct(MessageBusInterface $bus)
     {
-        $this->queryBus = $queryBus;
+        $this->bus = $bus;
+    }
+
+    protected function handle(Command $command): void
+    {
+        $this->bus->dispatch($command);
+    }
+
+    protected function handleWithDelay(Command $command): void
+    {
+        $this->bus->dispatch(new Envelope($command), [
+            new DelayStamp(840000),
+        ]);
     }
 
     protected function ask(Query $query): Envelope
     {
-        $envelope = $this->queryBus->dispatch($query);
+        $envelope = $this->bus->dispatch($query);
         /** @var HandledStamp $handled */
         $handled = $envelope->last(HandledStamp::class);
 
@@ -30,7 +43,7 @@ abstract class QueryController extends AbstractController
 
     protected function askWithDelay(Query $query): Envelope
     {
-        $envelope = $this->queryBus->dispatch(new Envelope($query), [
+        $envelope = $this->bus->dispatch(new Envelope($query), [
             new DelayStamp(840000),
         ]);
         /** @var HandledStamp $handled */
