--- src/Infrastructure/Doctrine/Repository/DoctrineProductRepository.php
@@ -1,17 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace App\Infrastructure\Doctrine\Repository;
-
-use App\DomainModel\Repository\ProductRepository;
-use App\Entity\Product;
-use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\Persistence\ManagerRegistry;
-
-final class DoctrineProductRepository extends ServiceEntityRepository implements ProductRepository
-{
-    public function __construct(ManagerRegistry $registry)
-    {
-        parent::__construct($registry, Product::class);
-    }
-}
